requirements.txt
Pillow
numpy
matplotlib
cython